Add handling of asks in gen_auction
This implementation assumes that the handling of duplicates should be
the same for bids and asks (configured via the {duplicate, Mode}
option). Handling them differently will be a pain, and I don't (or
don't want to) see the use case at the moment.

The only outstanding issue that came up while working on this was
whether or not to enforce mutual exclusion between bidders and
askers (i.e. a process that submits a bid cannot also submit an ask,
and vice versa). I'm leaving that for a future task. Similar to the
handling of different duplicate modes for bids and asks this seems
likely to be a pain, but I see the potential use case more clearly.