Fix error in duplicates test causing intermittent failure
This appears to have been the result of assuming that messages are sent by the
gen_auction process in a deterministic order (and are delivered in order);
however, it appears that this assumption is not true. They probably are
delivered in order, since Erlang guarantees message order between two processes.
That leaves a very mystifying behavior where the apparently deterministic
gen_auction module is somehow sending notification messages in a
nondeterministic order.